Basic WebSocket echo functionality now working
Can now get the server to send a client's message back through
WebSocket connection. Next step would be to modify the server to parse
and act on data as specified by a communication interface that will
have to be designed. (Consider JSON for defining the data)

Future consideration: will need to create embedded Glassfish server so
that it can be started up within the ISIS main application (while
hosting it's own application) that will be installed in the Innovation
Station.